initial pull request for isse 75, season endpoints
