+++ backend/internal/server/server.go
@@ -16,7 +16,7 @@ import (
 	_ "github.com/jak103/powerplay/internal/server/apis/notifications"
 	_ "github.com/jak103/powerplay/internal/server/apis/stats"
 	_ "github.com/jak103/powerplay/internal/server/apis/user"
-	_ "github.com/jak103/powerplay/internal/server/apis/season"
+	_ "github.com/jak103/powerplay/internal/server/apis/groups"
 
 )
