+++ static/oas/v1/openapi.yml
@@ -11,7 +11,7 @@ paths:
   /auth:
     $ref: "./auth/auth.yml#/paths/auth"
   /shotsongoal:
-    $ref: "./stats/shotsongoal.yml#/paths/shotsOnGoal"   
+    $ref: "./stats/shotsongoal.yml#/paths/shotsOnGoal"
   /penaltyTypes:
     $ref: "./stats/penalties.yml#/paths/penaltyTypes"
   /leagues:
@@ -25,6 +25,8 @@ paths:
   #$ref: "./[Relative path starting from v1]#/paths/[yml path]"
   /goals:
     $ref: "./stats/goal.yml#/paths/goals"
+  /seasons:
+    $ref: "./season/season.yml#/paths/seasons"
 
 components:
   securitySchemes:
