+++ static/oas/v1/season/season.yml
@@ -1,0 +1,149 @@
+paths:
+  seasons:
+    get:
+      tags:
+        - Season
+      summary: Get All Seasons
+      description: |
+        Get all seasons
+      responses:
+        200:
+          description: The response body should contain the list of seasons
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/GetSeasonsResponse"
+        400:
+          $ref: "../common/errors.yml#/responses/BadRequest"
+    post:
+      tags:
+        - Season
+      summary: Season post request
+      description: |
+        **REQUIRED PERMISSIONS:** none:none  
+        **RATE LIMIT:** TBD
+      requestBody:
+        description: The request body should contain a name, start time, end time, registrations, schedule, and leagues
+        required: true
+        content:
+          application/json:
+            schema:
+              $ref: "#/components/schemas/SeasonRequest"
+      responses:
+        200:
+          description: Successfully posted season
+          content:
+            application/json:
+              schema:
+                $ref: "#/components/schemas/SeasonResponse"
+        400:
+          $ref: "../common/errors.yml#/responses/BadRequest"
+
+components:
+  schemas:
+    GetSeasonsResponse:
+      type: object
+      properties:
+        status_code:
+          $ref: "../common/schemas.yml#/schemas/StatusCode200"
+        status_string:
+          $ref: "../common/schemas.yml#/schemas/StatusString200"
+        request_id:
+          $ref: "../common/schemas.yml#/schemas/RequestId"
+        response_data:
+          type: object
+          properties:
+            seasons:
+              type: array
+              description: All seasons
+              example:
+                - id: 1
+                  created_at: "2009-11-10T23:00:00Z"
+                  updated_at: "2009-11-10T23:00:00Z"
+                  name: "Season 1"
+                  start: "2009-11-10T23:00:00Z"
+                  end: "2009-11-10T23:00:00Z"
+                  registrations: []
+                  schedule: []
+                  leagues: []
+                - id: 2
+                  created_at: "2009-11-10T23:00:00Z"
+                  updated_at: "2009-11-10T23:00:00Z"
+                  name: "Season 2"
+                  start: "2009-11-10T23:00:00Z"
+                  end: "2009-11-10T23:00:00Z"
+                  registrations: []
+                  schedule: []
+                  leagues: []
+
+    SeasonRequest:
+      type: object
+      properties:
+        name:
+          type: string
+          description: The name of the season
+          example: "Season 1"
+        start:
+          type: string
+          format: date-time
+          description: Start time for the season
+          example: "2023-05-01T00:00:00Z"
+        end:
+          type: string
+          format: date-time
+          description: End time for the season
+          example: "2023-09-01T00:00:00Z"
+        registrations:
+          type: array
+          description: The registrations for the season
+          items:
+            type: object
+            example: {}
+          example: []
+        schedule:
+          type: array
+          description: The schedule for the season
+          items:
+            type: object
+            example: {}
+          example: []
+        leagues:
+          type: array
+          description: The leagues associated with the season
+          items:
+            type: object
+            properties:
+              correlation_id:
+                type: string
+                description: The correlation ID of the league
+                example: "abc-123-def"
+              season_id:
+                type: integer
+                description: The ID of the season
+                example: 1
+              name:
+                type: string
+                description: The name of the league
+                example: "Premier League"
+              teams:
+                type: array
+                description: The teams in the league
+                items:
+                  type: object
+          example:
+            - correlation_id: "abc-123-def"
+              season_id: 1
+              name: "Premier League"
+              teams: []
+
+    SeasonResponse:
+      type: object
+      properties:
+        status_code:
+          $ref: "../common/schemas.yml#/schemas/StatusCode200"
+        status_string:
+          $ref: "../common/schemas.yml#/schemas/StatusString200"
+        request_id:
+          $ref: "../common/schemas.yml#/schemas/RequestId"
+        response_data:
+          type: object
